break out foreign-key.ts from metamodel.ts and start initial break out module util.ts
